Add a test for the main program
